[Refactor]: 채팅 컴포넌트 분리
